Refactor dmsr_gan module into wgan and monitor modules
