+++ dmsr/monitors/__init__.py
@@ -1,0 +1,2 @@
+from .monitor import *
+from .manager import *
