+++ dmsr/wgan/__init__.py
@@ -2,4 +2,3 @@ from .wgan import *
 from .generator import *
 from .critic import *
 from .density_critic import *
-from .monitor import *
