+++ dmsr/wgan/blocks.py
