+++ dmsr/wgan/critic.py
