+++ dmsr/wgan/density_critic.py
