+++ dmsr/wgan/generator.py
