+++ dmsr/wgan/wgan.py
