+++ scripts/tests/dmsr_generator_test.py
@@ -13,7 +13,7 @@ sys.path.append("../..")
 import time
 import torch
 
-from dmsr.dmsr_gan import DMSRGenerator
+from dmsr.wgan import DMSRGenerator
 
 # Check if CUDA is available and set the device
 gpu_id = 0
